% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_s3.R
\name{fig_s3}
\alias{fig_s3}
\title{Generate Figure S3: Association between SIRI and Tumor Burden}
\usage{
fig_s3(data = pantheia::pantheia_data)
}
\arguments{
\item{data}{A dataframe. Defaults to \code{pantheia::pantheia_data}.}
}
\value{
A patchwork object.
}
\description{
Reproduces Supplementary Figure 3. This figure analyzes the association between 
log(SIRI) levels and various markers of tumor burden (Liver Metastases, Lung Metastases, 
Peritoneal Metastases, Liver Tumor Burden, VTE, and Obstructive Jaundice).
}
