% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pantheia_data}
\alias{pantheia_data}
\title{Clinical data from the PANTHEIA-SEOM study}
\format{
A data frame with one row per patient and the following variables:
\describe{
  \item{logsiri}{Natural logarithm of the SIRI index}
  \item{pfs_time1l}{Progression-free survival time (months)}
  ...
}
}
\source{
Internal PANTHEIA-SEOM study data.
}
\usage{
pantheia_data
}
\description{
Dataset containing clinical information and SIRI values for patients with
metastatic pancreatic ductal adenocarcinoma.
}
\keyword{datasets}
