% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Generate Baseline Patient Characteristics Table (Table 1)}
\usage{
table1(data = pantheia::pantheia_data)
}
\arguments{
\item{data}{A dataframe or a file path to a CSV. Defaults to \code{pantheia::pantheia_data}.}
}
\value{
A gtsummary object.
}
\description{
Reproduces Table 1. By default, it uses the internal package data (.rda) which preserves
factor levels (ordering). It can also accept a path to a CSV file.
}
