% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pantheia_data}
\alias{pantheia_data}
\title{Datos clínicos del estudio PanTheia}
\format{
Un data frame con filas y variables:
\describe{
  \item{logsiri}{Logaritmo del índice SIRI}
  \item{pfs_time1l}{Tiempo de supervivencia libre de progresión}
  ...
}
}
\source{
Datos internos del estudio.
}
\usage{
pantheia_data
}
\description{
Dataset conteniendo información clínica y valores SIRI de pacientes.
}
\keyword{datasets}
